Removed unnecessary entities from access-control service (Address, Campaign, Role, User)
