--- access-control/src/main/java/com/egovernment/accesscontrol/domain/entity/Address.java
@@ -1,19 +1,0 @@
-package com.egovernment.accesscontrol.domain.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Address {
-    private String country;
-    private String region;
-    private String municipality;
-    private String postcode;
-    private String city;
-    private String village;
-}
