--- access-control/src/main/java/com/egovernment/accesscontrol/domain/entity/Role.java
@@ -1,23 +1,0 @@
-package com.egovernment.accesscontrol.domain.entity;
-
-import com.egovernment.accesscontrol.enums.RoleEnum;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "roles")
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Role extends BaseEntity {
-
-    @Enumerated(EnumType.STRING)
-    @Column
-    private RoleEnum roleName;
-
-}
-
