+++ access-control/src/test/java/com/egovernment/accesscontrol/service/FilteredCampaignServiceTest.java
@@ -1,9 +1,9 @@
 package com.egovernment.accesscontrol.service;
 
-import com.egovernment.accesscontrol.domain.entity.Campaign;
+import com.egovernment.accesscontrol.domain.entity.FilteredCampaign;
+import com.egovernment.accesscontrol.repository.FilteredCampaignRepository;
 import com.egovernment.kafka.domain.dto.CampaignFilteredDTO;
 import com.egovernment.accesscontrol.enums.CampaignType;
-import com.egovernment.accesscontrol.repository.CampaignRepository;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
@@ -20,28 +20,28 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-public class CampaignServiceTest {
+public class FilteredCampaignServiceTest {
     @Mock
-    private CampaignRepository campaignRepository;
+    private FilteredCampaignRepository filteredCampaignRepository;
     @Mock
     private ModelMapper modelMapper;
 
     @InjectMocks
-    private CampaignService campaignServiceToTest;
+    private FilteredCampaignService filteredCampaignServiceToTest;
 
     private final String TEST_REGION_NAME = "TestRegion";
     private final String VOTING_CAMPAIGN_TYPE = "VOTING";
     private final String CENSUS_CAMPAIGN_TYPE = "CENSUS";
     private final String TEST_GLOBAL_REGION_NAME = "GLOBAL";
-    private final Campaign firstMockCampaign = Campaign.builder()
-            .campaignType(CampaignType.VOTING)
-            .title(VOTING_CAMPAIGN_TYPE)
+    private final FilteredCampaign firstMockCampaign = FilteredCampaign.builder()
+            .campaignType(String.valueOf(CampaignType.VOTING))
+            .campaignTitle(VOTING_CAMPAIGN_TYPE)
             .regionName(TEST_REGION_NAME)
             .build();
 
-    private final Campaign secondMockCampaign = Campaign.builder()
-            .campaignType(CampaignType.CENSUS)
-            .title(CENSUS_CAMPAIGN_TYPE)
+    private final FilteredCampaign secondMockCampaign = FilteredCampaign.builder()
+            .campaignType(String.valueOf(CampaignType.CENSUS))
+            .campaignTitle(CENSUS_CAMPAIGN_TYPE)
             .regionName(TEST_GLOBAL_REGION_NAME)
             .build();
     private final CampaignFilteredDTO firstMockCampaignDto = CampaignFilteredDTO.builder()
@@ -58,12 +58,12 @@ public class CampaignServiceTest {
     @Test
     public void testGetActiveLocalCampaigns() {
 
-        when(campaignRepository.findAll()).thenReturn(Arrays.asList(firstMockCampaign, secondMockCampaign));
+        when(filteredCampaignRepository.findAll()).thenReturn(Arrays.asList(firstMockCampaign, secondMockCampaign));
 
         lenient().when(modelMapper.map(eq(firstMockCampaign), eq(CampaignFilteredDTO.class))).thenReturn(firstMockCampaignDto);
         lenient().when(modelMapper.map(eq(secondMockCampaign), eq(CampaignFilteredDTO.class))).thenReturn(secondMockCampaignDto);
 
-        List<CampaignFilteredDTO> result = campaignServiceToTest.getActiveLocalCampaigns(TEST_REGION_NAME);
+        List<CampaignFilteredDTO> result = filteredCampaignServiceToTest.getActiveLocalCampaigns(TEST_REGION_NAME);
 
         assertEquals(2, result.size());
